// frontend/src/api/generated/api.ts
export interface UserCreateRequestDTO {
    'name'?: string;
    'surname'?: string;
    'password'?: string;
    'username'?: string;
    'email'?: string;
    'phoneNumber'?: string;
}